+++ opsgenie/resource_opsgenie_schedule.go
@@ -164,7 +164,7 @@ func resourceOpsgenieScheduleDelete(d *schema.ResourceData, meta interface{}) er
 
 func validateOpsgenieScheduleName(v interface{}, k string) (ws []string, errors []error) {
 	value := v.(string)
-	if !regexp.MustCompile(`^[a-zA-Z0-9_-]+$`).MatchString(value) {
+	if !regexp.MustCompile(`^[a-zA-Z 0-9_-]+$`).MatchString(value) {
 		errors = append(errors, fmt.Errorf(
 			"only alpha numeric characters and underscores are allowed in %q: %q", k, value))
 	}